# Conflicts:
#	Jiggie/src/main/java/com/jiggie/android/component/Utils.java

+++ Jiggie/src/main/java/com/jiggie/android/activity/event/EventDetailActivity.java
@@ -31,6 +31,7 @@ import android.widget.Toast;
 
 import com.jiggie.android.App;
 import com.jiggie.android.R;
+import com.jiggie.android.activity.ecommerce.ProductListActivity;
 import com.jiggie.android.component.FlowLayout;
 import com.jiggie.android.component.StringUtility;
 import com.jiggie.android.component.Utils;
@@ -154,7 +155,6 @@ public class EventDetailActivity extends ToolbarActivity implements SwipeRefresh
                     e.printStackTrace();
                 }
             }
-
         }
 
         if(event_name!=null){
@@ -177,9 +177,7 @@ public class EventDetailActivity extends ToolbarActivity implements SwipeRefresh
 
         super.registerReceiver(this.guestInvitedReceiver, new IntentFilter(super.getString(R.string.broadcastGuestInvited)));
         App.getInstance().trackMixPanelEvent("View Event Details");
-
-        file = new File(Environment.getExternalStorageDirectory() + "event.png");
-        if(file.exists())
+        if(file!= null && file.exists())
             file.delete();
     }
 
@@ -274,20 +272,20 @@ public class EventDetailActivity extends ToolbarActivity implements SwipeRefresh
                     }
 
                     //not finished yet----
-                /*for (int i = guestCount; i > 0; i--) {
-                    final String url = App.getFacebookImage(guestArr.get(guestCount - i).getFb_id(), width);
-                    Glide.with(EventDetailActivity.this).load(url).asBitmap().centerCrop().into(new BitmapImageViewTarget(imageGuests[i]) {
-                        @Override
-                        protected void setResource(Bitmap resource) {
-                            final Resources resources = getResources();
-                            if (resources != null) {
-                                final RoundedBitmapDrawable circularBitmapDrawable = RoundedBitmapDrawableFactory.create(resources, resource);
-                                circularBitmapDrawable.setCircular(true);
-                                super.getView().setImageDrawable(circularBitmapDrawable);
+                    /*for (int i = guestCount; i > 0; i--) {
+                        final String url = App.getFacebookImage(guestArr.get(guestCount - i).getFb_id(), width);
+                        Glide.with(EventDetailActivity.this).load(url).asBitmap().centerCrop().into(new BitmapImageViewTarget(imageGuests[i]) {
+                            @Override
+                            protected void setResource(Bitmap resource) {
+                                final Resources resources = getResources();
+                                if (resources != null) {
+                                    final RoundedBitmapDrawable circularBitmapDrawable = RoundedBitmapDrawableFactory.create(resources, resource);
+                                    circularBitmapDrawable.setCircular(true);
+                                    super.getView().setImageDrawable(circularBitmapDrawable);
+                                }
                             }
-                        }
-                    });
-                }*/
+                        });
+                    }*/
                     //-----------------------
                 }
 
@@ -315,17 +313,34 @@ public class EventDetailActivity extends ToolbarActivity implements SwipeRefresh
                 invalidateOptionsMenu();
                 populateTags();
 
-                if (StringUtility.isEquals(EventManager.FullfillmentTypes.PHONE_NUMBER, message.getData().getEvents_detail().getFullfillment_type(), true)) {
+                final String fullfillmentType = message.getData().getEvents_detail().getFullfillment_type();
+                btnBook.setVisibility(
+                        StringUtility.isEquals(EventManager.FullfillmentTypes.NONE, message.getData().getEvents_detail().getFullfillment_type(), true)
+                                ? View.GONE : View.VISIBLE);
+                /*if(EventManager.FullfillmentTypes.NONE.equals(fullfillmentType))
+                {
+                    btnBook.setVisibility(View.GONE);
+                }
+                else if(EventManager.FullfillmentTypes.TICKET.equals(fullfillmentType))
+                {
+                    btnBook.setVisibility(View.VISIBLE);
+                }
+                else btnBook.setVisibility(View.VISIBLE);*/
+
+
+                if (StringUtility.isEquals(EventManager.FullfillmentTypes.PHONE_NUMBER, fullfillmentType, true)) {
                     txtExternalSite.setVisibility(View.GONE);
                     txtBookNow.setText(R.string.call);
-                } else if (StringUtility.isEquals(EventManager.FullfillmentTypes.RESERVATION, message.getData().getEvents_detail().getFullfillment_type(), true)) {
+                } else if (StringUtility.isEquals(EventManager.FullfillmentTypes.RESERVATION, fullfillmentType, true)) {
                     txtExternalSite.setVisibility(View.GONE);
                     txtBookNow.setText(R.string.reserve);
-                } else if (StringUtility.isEquals(EventManager.FullfillmentTypes.PURCHASE, message.getData().getEvents_detail().getFullfillment_type(), true)) {
+                } else if (StringUtility.isEquals(EventManager.FullfillmentTypes.PURCHASE, fullfillmentType, true)) {
                     txtExternalSite.setVisibility(View.GONE);
                     txtBookNow.setText(R.string.purchase);
+                } else if (StringUtility.isEquals(EventManager.FullfillmentTypes.TICKET, fullfillmentType, true)) {
+                    txtExternalSite.setVisibility(View.GONE);
+                    txtBookNow.setText(getResources().getString(R.string.purchase_ticket));
                 }
-
                 //Changed by Aga 16-2-2016
                 else if (StringUtility.isEquals(EventManager.FullfillmentTypes.LINK, message.getData().getEvents_detail().getFullfillment_type(), true)) {
                     txtExternalSite.setVisibility(View.VISIBLE);
@@ -438,6 +453,12 @@ public class EventDetailActivity extends ToolbarActivity implements SwipeRefresh
 
             }else if (StringUtility.isEquals(EventManager.FullfillmentTypes.NONE, this.eventDetail.getFullfillment_type(), true))
                 Toast.makeText(this, R.string.no_fullfillment, Toast.LENGTH_SHORT).show();
+            else if(StringUtility.isEquals(EventManager.FullfillmentTypes.TICKET, this.eventDetail.getFullfillment_type(), true))
+            {
+                Intent i = new Intent(this, ProductListActivity.class);
+                i.putExtra(Common.FIELD_EVENT_ID, this.eventDetail.get_id());
+                startActivity(i);
+            }
             else
                 Toast.makeText(this, R.string.book_error, Toast.LENGTH_SHORT).show();
         }
